moved views to own modules
