+++ commit-scoreboard/scoreboard/adapters/view_debugger.py
@@ -1,0 +1,10 @@
+from pprint import pprint
+
+from scoreboard.core.app import AppModel, ScoreboardView
+
+
+
+class ConsoleView(ScoreboardView):
+    def update(self, model: AppModel) -> None:
+        pprint(dict(model.statuses))
+
